requirements.txt
ushuffle~=1.1.2
biopython~=1.79
h5py~=3.1.0
jupyterlab~=3.0.16  
keras==2.4.3
numpy~=1.19.2
pandas~=1.2.5
pytest~=6.2.4  
scikit-learn~=0.24.2
scipy~=1.7.0
seaborn~=0.11.1
seqlogo~=5.29.8
setuptools~=57.0.0
statsmodels~=0.12.2
tensorflow==2.5.0
tables~=3.6.1
